# backend/internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: Framely HTTP API
  version: "1.0.0"
  description: |
    路径前缀是 `/api`。前后端一起发布，不按路径做 API 版本。

    业务 JSON 信封为 `{ code, data, msg, reason }`。`code === 0` 表示成功；失败时 HTTP status 与
    `code` 都表达失败。`code` 默认等于 HTTP 状态；同一状态下的细分原因使用 `status*100+n`
    （例如 40301 配额、42901 限流）。机器判断请用 `reason`，不要解析 `msg`。

    TypeScript 调用面仍是 `web/src/services/api` 的 `http` 模块，不从本规范生成客户端。
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
  /account/login:
    post:
      tags: [gateway]
      summary: 账号登录并签发访问密钥
      description: 本地后端代为调用部署配置的 new-api 兼容服务（BEEFTV_GATEWAY_BASE_URL）完成登录、双重验证和令牌签发；复用或创建 beeftv-auto 前缀的密钥。服务端不保存会话，调用方把返回的密钥写入账号模型渠道。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
                twoFactorCode: { type: string }
      responses:
        "200":
          description: 信封 data 为 { baseUrl, apiKey, keyName, account }
        "400":
          description: 账号或验证码错误；reason 为 two_factor_required 时需补充双重验证码
        "502":
          description: 模型服务不可达或返回异常
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: open_ai_canvas_session
...